pyproject.toml
[tool.poetry]
name = "sokora"
version = "0.1.0"
description = "勤怠管理のためのシンプルなWebアプリケーション"
authors = ["Sokora Team <info@example.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.103.1"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
jinja2 = "^3.1.2"
python-multipart = "^0.0.6"
pydantic = "^2.4.0"
sqlalchemy = "^2.0.23"
jpholiday = "^1.0.1"
pytz = "^2023.3"
python-dateutil = "^2.9.0.post0"
aiosqlite = "^0.20.0"
types-python-dateutil = "^2.9.0.20241206"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
isort = "^5.12.0"
mypy = "^1.10.0"
pytest = "^8.2.2"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.23.7"
httpx = "^0.27.0"
ruff = "^0.5.0"
pytest-playwright = "^0.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py313"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.13"
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
# 必要に応じて他の pytest オプションを追加
# asyncio_mode = "strict"
asyncio_mode = "auto"
